# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Document Recognition API",
        "description": "API для распознавания документов через внешний сервис bescan.ru",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8080/api/v1",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/recognize": {
            "post": {
                "tags": [
                    "Document Recognition"
                ],
                "summary": "Отправить документ на распознавание",
                "description": "Отправляет документ на распознавание во внешний сервис bescan.ru",
                "operationId": "recognizeDocument",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "document_id",
                                    "document_type",
                                    "images"
                                ],
                                "properties": {
                                    "document_id": {
                                        "description": "Уникальный идентификатор документа",
                                        "type": "string",
                                        "example": "doc_123"
                                    },
                                    "document_type": {
                                        "description": "Тип документа",
                                        "type": "string",
                                        "enum": [
                                            "PASSPORT",
                                            "PASSPORT_REG",
                                            "DLIC",
                                            "SNILS",
                                            "STS"
                                        ],
                                        "example": "PASSPORT"
                                    },
                                    "images": {
                                        "description": "Массив изображений в формате base64",
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "base64"
                                        },
                                        "example": [
                                            "/9j/4AAQSkZJRgABAQEASABIAAD/4gIYSUNDX1BST0ZJTEUAAQEAAAIIAAAAAAQwAABtbnRyUkdCIFhZWiAH"
                                        ]
                                    },
                                    "callback_url": {
                                        "description": "URL для отправки результата распознавания",
                                        "type": "string",
                                        "format": "url",
                                        "example": "https://example.com/webhook"
                                    },
                                    "metadata": {
                                        "description": "Дополнительные метаданные",
                                        "type": "object",
                                        "example": {
                                            "user_id": 123,
                                            "session_id": "abc123"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Документ успешно отправлен на распознавание",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "task_id": {
                                            "description": "ID задачи в системе",
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "document_id": {
                                            "description": "ID документа (дублирует task_id)",
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "external_task_id": {
                                            "description": "ID задачи во внешнем сервисе",
                                            "type": "string",
                                            "example": "s-12345"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Document sent for recognition"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка при отправке документа",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "External API error: Connection timeout"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "document_type": [
                                                    "The document type field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/status/{task_id}": {
            "get": {
                "tags": [
                    "Document Recognition"
                ],
                "summary": "Получить статус задачи распознавания",
                "description": "Возвращает текущий статус задачи распознавания и результат, если готов",
                "operationId": "getTaskStatus",
                "parameters": [
                    {
                        "name": "task_id",
                        "in": "path",
                        "description": "ID задачи, внешний ID задачи или ID документа",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Статус задачи получен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "task_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "document_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "external_task_id": {
                                            "type": "string",
                                            "example": "s-12345"
                                        },
                                        "status": {
                                            "type": "string",
                                            "enum": [
                                                "pending",
                                                "processing",
                                                "completed",
                                                "failed"
                                            ],
                                            "example": "completed"
                                        },
                                        "document_type": {
                                            "type": "string",
                                            "example": "PASSPORT"
                                        },
                                        "metadata": {
                                            "type": "object",
                                            "example": {
                                                "user_id": 123
                                            }
                                        },
                                        "data": {
                                            "description": "Распознанные данные (только для completed)",
                                            "type": "object",
                                            "example": {
                                                "Series": "1234",
                                                "Number": "567890"
                                            }
                                        },
                                        "confidences": {
                                            "description": "Уверенность в распознавании (только для completed)",
                                            "type": "object",
                                            "example": {
                                                "Series": 0.95
                                            }
                                        },
                                        "verifications": {
                                            "description": "Результаты верификации (только для completed)",
                                            "type": "object",
                                            "example": {
                                                "Series": {
                                                    "valid": true
                                                }
                                            }
                                        },
                                        "error": {
                                            "description": "Описание ошибки (только для failed)",
                                            "type": "string",
                                            "example": "Document processing failed"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Задача не найдена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Task not found"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 404
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}